issue11 html layout of the appointment main view.
Created the appointment main view in www/templates/appointments.html.

The main view contains a home button as part of the header, a left and a right button to allow the user go to the previous/next month, the text in between of the left and the right button indicates the current month. The calender is implemented though a 6 by 7 table.

Todo: Not sure where the home button should link to.