Created Measurement History view in HTML
Whenever a user clicks measurements tab they will be directed to the history view which will load a list of there previous meausurements.

The measurements appear in this form
Date
Measurement 1
Measurement 2
...etc

I did not create a nicer view for the measurement list because, the app should send the user to apples HealthKit to view all thier measurements.