issue15 change background of today's date
Added the css file of appointment into index.html header.
Modified controller to add a variable of current date.
Add ng-class style attribute to tr in html file. So if the iterated date muches current date will change the css style. Which is has a blue border around that tr.