Displayed a popup notification for invalid blood pressure
I use an ng.blur event to trigger a function to check if the blood
pressure reading is too high or low (just systolic). I hardcoded
limit values for now. I also changed the color if the reading is
invalid. I alos changed the measurements css file so the measurement
labels have an auto margin.