trying to use ui-sref but failed.
There is no way to incorperate a conditional ui-sref link in this case. Tried ui-sref-active, eat-click-if, ng -if, incorporate condition into ui-sref statement but nothing worked. The best result I can get is current version which has a bug that if the user clicked a date without event, it will enter an empty event detail page.

I highly suggest use the previous version where in the controller do the judgement of whether the date is an event day and only redirect to the detailed appointment view if it is an event day.