Implemented the html of input a medication by manually input.
Problem:
I'm confused by the medication name and medication trad_name  as defined in the medication factory. Also Based on our wireframe, there are more attributes than those defined in the medication factory. For instance, whether this med is a sheduled med or as needed med.

To view the result, type #/input_main into your url